[FEAT/#23] RemoveBackground 버튼 누르면 바로 배경이 제거된 이미지로 나타나게 구현
