+++ BiRTH/BiRTH/View/Soy/RemoveImageBackgroundSheet.swift
@@ -6,6 +6,9 @@
 //
 
 import SwiftUI
+import Vision
+import CoreImage.CIFilterBuiltins
+
 
 struct RemoveImageBackgroundSheet: View {
     let rows = [GridItem(.flexible())]
@@ -13,7 +16,8 @@ struct RemoveImageBackgroundSheet: View {
     @Environment(\.dismiss) var dismiss
     
     @Binding var image: UIImage
-    @State var showRemoveImgBg = false
+    
+    var processingQueue = DispatchQueue(label: "ProcessingQueue")
     
     var body: some View {
         VStack {
@@ -29,7 +33,7 @@ struct RemoveImageBackgroundSheet: View {
             ScrollView(.horizontal) {
                 LazyHGrid(rows: rows) {
                     Button {
-                        showRemoveImgBg.toggle()
+                        createSticker()
                         print("removeImageBackground")
                     } label: {
                         ZStack {
@@ -42,9 +46,6 @@ struct RemoveImageBackgroundSheet: View {
                                 
                         }
                     }
-                    .sheet(isPresented: $showRemoveImgBg) {
-                        RemoveImageBackground(image: $image)
-                    }
                 }
             }
         }
@@ -52,6 +53,65 @@ struct RemoveImageBackgroundSheet: View {
     }
 }
 
-//#Preview {
-//    RemoveImageBackgroundSheet(image: <#UIImage#>)
-//}
+
+extension RemoveImageBackgroundSheet {
+    func createSticker() {
+        guard let inputImage = CIImage(image: image) else {
+            print("Failed to create CIImage")
+            return
+        }
+        
+        processingQueue.async {
+            guard let maskImage = self.subjectMaskImage(from: inputImage) else {
+                print("Failed to create mask image")
+                return
+            }
+            
+            let outputImage = self.apply(mask: maskImage, to: inputImage)
+            let image = self.render(ciImage: outputImage)
+            DispatchQueue.main.async {
+                self.image = image
+            }
+        }
+    }
+    
+    
+    func subjectMaskImage(from inputImage: CIImage) -> CIImage? {
+        let handler = VNImageRequestHandler(ciImage: inputImage)
+        let request = VNGenerateForegroundInstanceMaskRequest()
+        
+        do {
+            try handler.perform([request])
+            
+            guard let result = request.results?.first else {
+                print("No observations found")
+                return nil
+            }
+            
+            do {
+                let maskPixelBuffler = try result.generateScaledMaskForImage(forInstances: result.allInstances, from: handler)
+                return CIImage(cvImageBuffer: maskPixelBuffler)
+            }
+        } catch {
+            print(error)
+            return nil
+        }
+    }
+    
+    func apply(mask: CIImage, to image: CIImage) -> CIImage {
+        let filter = CIFilter.blendWithMask()
+        filter.inputImage = image
+        filter.maskImage = mask
+        filter.backgroundImage = CIImage.empty()
+        
+        return filter.outputImage!
+    }
+    
+    func render(ciImage: CIImage) -> UIImage {
+        guard let cgImage = CIContext(options: nil).createCGImage(ciImage, from: ciImage.extent) else {
+            fatalError("Failed to render CGImage")
+        }
+        return UIImage(cgImage: cgImage)
+    }
+}
+
